[APTR-36/37] Add home page
