+++ .expo/types/router.d.ts
@@ -7,7 +7,7 @@ declare module "expo-router" {
   export * from 'expo-router/build';
 
   // prettier-ignore
-  type StaticRoutes = `/`;
+  type StaticRoutes = `/` | `/auth/`;
   // prettier-ignore
   type DynamicRoutes<T extends string> = never;
   // prettier-ignore
