+++ src/app/index.tsx
@@ -1,78 +1,23 @@
-import { Button, SafeAreaView, StyleSheet, View } from "react-native";
-import { ClerkProvider, SignedIn, SignedOut, useAuth } from "@clerk/clerk-expo";
-import SignInWithOAuth from "../components/SignInWithOAuth";
-import FetchUser from "../components/FetchUser";
-import * as SecureStore from "expo-secure-store";
-
-const CLERK_PUBLISHABLE_KEY = process.env.EXPO_PUBLIC_CLERK_PUBLISHABLE_KEY!;
-
-const tokenCache = {
-  async getToken(key: string) {
-    try {
-      return SecureStore.getItemAsync(key);
-    } catch (err) {
-      return null;
-    }
-  },
-  async saveToken(key: string, value: string) {
-    try {
-      return SecureStore.setItemAsync(key, value);
-    } catch (err) {
-      return;
-    }
-  },
-};
-
-const SignOut = () => {
-  const { isLoaded, signOut } = useAuth();
-  if (!isLoaded) {
-    return null;
-  }
-  return (
-    <View>
-      <Button 
-        title="Sign Out"
-        onPress={() => {
-          signOut();
-        }}
-      />
-    </View>
-  );
-};
+import { SafeAreaView, Text, Button, StyleSheet } from 'react-native';
+import { useRouter } from 'expo-router';
 
 export default function App() {
-  return (
-    <ClerkProvider
-      tokenCache={tokenCache}
-      publishableKey={CLERK_PUBLISHABLE_KEY}
-    >
-      <SafeAreaView style={styles.container}>
-        <SignedIn>
-          <FetchUser />
-          <SignOut />
-        </SignedIn>
-        <SignedOut>
-          <SignInWithOAuth />
-        </SignedOut>
-      </SafeAreaView>
-    </ClerkProvider>
-  );
-}
+
+    const navigation = useRouter();
+
+    return (
+    <SafeAreaView style={styles.container}>
+        <Text>Welcome To Off Campus!</Text>
+        <Button title="Sign In" onPress={() => navigation.push("/auth" as any)} />
+    </SafeAreaView>
+    );
+};
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  buttonContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: "#000", 
-    borderRadius: 4,
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-  },
-});
+    container: {
+      flex: 1,
+      backgroundColor: "#fff",
+      alignItems: "center",
+      justifyContent: "center",
+    }
+  });
